refactor: rename to objekt-stats
