+++ README.md
@@ -1,6 +1,6 @@
-# Cosmo Stats
+# Objekt Stats
 
-Cosmo Stats is a command line tool that retrieves the total no. of Objekt minted
+Objekt Stats is a command line tool that retrieves the total no. of Objekt minted
 for members of a MODHAUS group.
 
 ## Installation
@@ -10,29 +10,29 @@ Installation via [pipx](https://pipx.pypa.io/stable/) is recommended.
 Via Git repository:
 
 ```sh
-pipx install git+https://github.com/hillaryychan/cosmo-stats.git@main
+pipx install git+https://github.com/hillaryychan/objekt-stats.git@main
 ```
 
-Check that Cosmo Stats is available by running the `cosmo-stats` command:
+Check that Objekt Stats is available by running the `objekt-stats` command:
 
 ```sh
-cosmo-stats --help
+objekt-stats --help
 ```
 
 ## Usage
 
-The basic usage of `cosmo-stats` requires the artist, season and collection numbers of the objekts:
+The basic usage of `objekt-stats` requires the artist, season and collection numbers of the objekts:
 
 ```sh
-cosmo-stats tripleS atom02 --collection-no 101z,102z,103z,104z,105z,106z,107z,108z
+objekt-stats tripleS atom02 --collection-no 101z,102z,103z,104z,105z,106z,107z,108z
 ```
 
 Each season has three editions. The collection numbers remain constant for each edition, so there is
 an alias argument that can be used when retrieving data for Objekts from editions:
 
 ```sh
-cosmo-stats tripleS atom02 --edition 1
+objekt-stats tripleS atom02 --edition 1
 ```
 
-If you do not provide a collection number or edition, `cosmo-stats` will retrieve data for **all**
+If you do not provide a collection number or edition, `objekt-stats` will retrieve data for **all**
 Objekts for the season. This is not recommended as you may get rate limited.
