+++ src/objekt_stats/enums.py
