+++ src/objekt_stats/objekts/models.py
@@ -1,6 +1,6 @@
 from pydantic import BaseModel
 
-from cosmo_stats.enums import ArtmsMember, Season, TripleSMember
+from objekt_stats.enums import ArtmsMember, Season, TripleSMember
 
 
 class ObjektCollectionData(BaseModel):
